Refactored test_registration_widget.py to account for funcitons moving to utils
