+++ .gitignore
@@ -85,4 +85,3 @@ venv/
 
 # Other test files
 Logs/
-rand_test.ipynb
